Part 1/requirements.txt
langchain-community>=0.2.0
langgraph>=0.0.69
ipython>=8.0.0
typing-extensions>=4.12.0
langchain-core>=0.2.0
pydantic>=2.0.0
requests>=2.32.0
langchain-google-genai>=1.0.6
python-dotenv>=1.0.0